requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-crispy-forms==1.14.0
django-redis==5.2.0
django_debug_toolbar==3.8.1
executing==1.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.5.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
